Rename pathfinder compatibility checks module.
Use a more explicit module name for the compatibility-check API and keep imports aligned with the renamed implementation.

Made-with: Cursor
