+++ cuda_pathfinder/cuda/pathfinder/_compatibility_checks.py
